requirements.txt
functions-framework==3.0.0
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
google-api-core==2.8.0
google-auth==2.6.6
google-cloud-bigquery==3.1.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.1
grpcio==1.46.3
grpcio-status==1.46.3
idna==3.3
numpy==1.22.4
packaging==21.3
proto-plus==1.20.4
protobuf==3.20.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9